Hsiao-Ying Lin and Wen-Guey Tzeng/webApp/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
gunicorn==20.0.4
idna==2.10
Naked==0.1.31
numpy
pycparser==2.20
pycryptodome==3.10.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
sqlparse==0.4.1
urllib3==1.26.3
